Implement force_delete

+++ hsdp/resource_dicom_object_store.go
@@ -17,6 +17,7 @@ func resourceDICOMObjectStore() *schema.Resource {
 		CreateContext: resourceDICOMObjectStoreCreate,
 		ReadContext:   resourceDICOMObjectStoreRead,
 		DeleteContext: resourceDICOMObjectStoreDelete,
+		SchemaVersion: 1,
 
 		Schema: map[string]*schema.Schema{
 			"config_url": {
@@ -32,6 +33,12 @@ func resourceDICOMObjectStore() *schema.Resource {
 			"description": {
 				Type:     schema.TypeString,
 				Optional: true,
+				ForceNew: true,
+			},
+			"force_delete": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  false,
 				ForceNew: true,
 			},
 			"static_access": {
@@ -133,11 +140,16 @@ func resourceDICOMObjectStoreDelete(_ context.Context, d *schema.ResourceData, m
 	config := m.(*Config)
 	configURL := d.Get("config_url").(string)
 	orgID := d.Get("organization_id").(string)
+	forceDelete := d.Get("force_delete").(bool)
 	client, err := config.getDICOMConfigClient(configURL)
 	if err != nil {
 		return diag.FromErr(err)
 	}
 	defer client.Close()
+	if !forceDelete { // soft delete
+		d.SetId("")
+		return diags
+	}
 	operation := func() error {
 		var resp *dicom.Response
 		_, resp, err = client.Config.DeleteObjectStore(dicom.ObjectStore{ID: d.Id()}, &dicom.QueryOptions{
